test: Delete HookRunners.test.js
The hooks.test.js and hooks-async.test.js files already cover all of the same test cases and are easier to maintain since they don't require mocking the request and reply objects.

(I realize those unit tests are difficult to maintain because of high coupling between the hook runners and the request/reply object being passed to them, but the implementation does simplify the code and reduce the depth of function calls. The integration-style tests cover everything anyway so it's not so bad.)