Remove res.notFound()
This method is too brittle. Not-found handlers are set for either the root app or sub-apps with a prefix. When a handler in any sub-app can invoke this method, it becomes unclear what the RouteContext of the not-found handler should be when that happens. Furthermore, there are potential bugs if this method is called inside an async hook (i.e. the hook resolves and continues the lifecycle even though the not-found handler was invoked).